waitingList, runningListと関連する操作を別ファイルに移す
